+++ src/main/kotlin/com/teamsparta/moamoa/domain/like/dto/LikeResponse.kt
