+++ src/main/kotlin/com/teamsparta/moamoa/domain/like/repository/LikeRepository.kt
