+++ src/main/kotlin/com/teamsparta/moamoa/domain/product/repository/ProductRepository.kt
@@ -1,0 +1,13 @@
+package com.teamsparta.moamoa.domain.product.repository
+
+import com.teamsparta.moamoa.domain.product.model.Product
+import org.springframework.data.jpa.repository.JpaRepository
+import java.util.*
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
+
+interface ProductRepository : JpaRepository<Product, Long>, CustomProductRepository {
+    fun findByIdAndDeletedAtIsNull(productId: Long): Optional<Product>
+    fun findAllByDeletedAtIsNull(pageable: Pageable): Page<Product>
+}
+
