+++ src/main/kotlin/com/teamsparta/moamoa/domain/review/dto/UpdateReviewRequest.kt
@@ -1,0 +1,17 @@
+package com.teamsparta.moamoa.domain.review.dto
+
+import com.teamsparta.moamoa.domain.review.model.Review
+
+class UpdateReviewRequest(
+    val title: String,
+    val content: String,
+    val name: String,
+    val imageUrl: String,
+) {
+    fun toUpdateReview(review: Review) {
+        review.title = this.title
+        review.content = this.content
+        review.name = this.name
+        review.imageUrl = this.imageUrl
+    }
+}
