+++ src/main/kotlin/com/teamsparta/moamoa/domain/user/repository/UserRepository.kt
@@ -1,0 +1,10 @@
+package com.teamsparta.moamoa.domain.user.repository
+
+import com.teamsparta.moamoa.domain.user.model.User
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserRepository : JpaRepository<User, Long> {
+    fun existsByEmail(email: String): Boolean
+
+    fun findByEmail(email: String): User?
+}
