--- src/main/kotlin/com/teamsparta/moamoa/product/repository/ProductStockRepository.kt
@@ -1,7 +1,0 @@
-package com.teamsparta.moamoa.product.repository
-
-import com.teamsparta.moamoa.product.model.ProductStock
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface ProductStockRepository :
-    JpaRepository<ProductStock, Long>
