--- src/main/kotlin/com/teamsparta/moamoa/user/model/User.kt
@@ -1,27 +1,0 @@
-package com.teamsparta.moamoa.user.model
-
-import com.teamsparta.moamoa.infra.BaseTimeEntity
-import jakarta.persistence.*
-import java.time.LocalDateTime
-
-@Table(name = "app_user")
-@Entity
-class User(
-    @Column(name = "email")
-    var email: String,
-    @Column(name = "password")
-    var password: String,
-    @Column(name = "nickname")
-    var nickname: String,
-    @Column(name = "address")
-    var address: String,
-    @Column(name = "phone_number")
-    var phoneNumber: String,
-) : BaseTimeEntity() {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
-
-    @Column(name = "deleted_at")
-    var deletedAt: LocalDateTime? = null
-}
