--- src/main/kotlin/com/teamsparta/moamoa/user/repository/UserRepository.kt
@@ -1,10 +1,0 @@
-package com.teamsparta.moamoa.user.repository
-
-import com.teamsparta.moamoa.user.model.User
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserRepository : JpaRepository<User, Long> {
-    fun existsByEmail(email: String): Boolean
-
-    fun findByEmail(email: String): User?
-}
